refactored directories
